refactor: separate out course navebar client component
