--- src/lib/store/atom.ts
@@ -1,19 +1,0 @@
-import { atom } from 'jotai';
-
-export interface CoursePageNavigationTree {
-  type: 'folder';
-  name: string;
-  children?: CoursePageNavigationTreeItem[];
-}
-
-export interface CoursePageNavigationTreeItem {
-  type: 'page' | 'folder' | 'separator';
-  name: string;
-  url: string;
-  external?: boolean;
-}
-
-export const coursePageNavigationAtom = atom<CoursePageNavigationTree | null>(
-  null,
-);
-coursePageNavigationAtom.debugLabel = 'coursePageNavigationAtom';
